% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{shinyInput}
\alias{shinyInput}
\title{}
\usage{
shinyInput(FUN, n, id, ...)
}
\arguments{
\item{FUN}{function to create the input}

\item{n}{number of inputs to be created}

\item{id}{ID prefix for each input}
}
\description{
md <- tempfile(fileext = '.md')
}
\details{

}
